# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organizations Directory Api",
        "description": "API for get information about organizations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://webserver:7020",
            "description": "local server"
        }
    ],
    "paths": {
...
        "/api/organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get all organizations with optional filtering",
                "operationId": "a6f3e52e9b7995087fc40990424c2977",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "building",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrganizationResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/organizations/{organization}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organization by ID",
                "operationId": "effcac3de5c029c2372022990ad7a48a",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organization not found"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/organizations/in-radius": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organizations within radius from coordinates",
                "operationId": "67a03cacbe0dc3e9cb6bd1e5531fec6a",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "building",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrganizationResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/organizations/in-rectangle": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organizations within rectangular area",
                "operationId": "f9cf6464a477e6d24faa1e6a5cb7fe48",
                "parameters": [
                    {
                        "name": "north",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "south",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "east",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "west",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "building",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrganizationResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid API key"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ActivityResource": {
                "title": "Activity Resource",
                "properties": {
                    "id": {
                        "description": "ID вида деятельности",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название вида деятельности",
                        "type": "string"
                    },
                    "parent": {
                        "$ref": "#/components/schemas/ActivityResource"
                    },
                    "children": {
                        "description": "Дочерние виды деятельности",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActivityResource"
                        }
                    }
                },
                "type": "object"
            },
            "BuildingResource": {
                "title": "Building Resource",
                "properties": {
                    "id": {
                        "description": "ID здания",
                        "type": "integer"
                    },
                    "address": {
                        "description": "Адрес здания",
                        "properties": {
                            "city": {
                                "type": "string"
                            },
                            "street": {
                                "type": "string"
                            },
                            "house": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "latitude": {
                        "description": "Географическая широта",
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "description": "Географическая долгота",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "OrganizationPhoneResource": {
                "title": "Organization Phone Resource",
                "properties": {
                    "id": {
                        "description": "ID телефона",
                        "type": "integer"
                    },
                    "number": {
                        "description": "Номер телефона",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "OrganizationResource": {
                "title": "Organization Resource",
                "properties": {
                    "id": {
                        "description": "ID организации",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название организации",
                        "type": "string"
                    },
                    "building": {
                        "$ref": "#/components/schemas/BuildingResource"
                    },
                    "phones": {
                        "description": "Телефоны организации",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrganizationPhoneResource"
                        }
                    },
                    "activities": {
                        "description": "Виды деятельности организации",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActivityResource"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}